requirements.txt
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.8.0
AHRS==0.3.1
ament-cmake-test==1.3.11
ament-copyright==0.12.12
ament-cppcheck==0.12.12
ament-cpplint==0.12.12
ament-flake8==0.12.12
ament-index-python==1.4.0
ament-lint==0.12.12
ament-lint-cmake==0.12.12
ament-package==0.14.0
ament-pep257==0.12.12
ament-uncrustify==0.12.12
ament-xmllint==0.12.12
angles==1.15.0
appdirs==1.4.4
apturl==0.5.2
argcomplete==3.6.0
attrs==25.1.0
backports.tarfile==1.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
beniget==0.4.2.post1
blinker==1.9.0
bloom==0.12.0
breezy==3.2.1
Brlapi==0.8.3
Brotli==1.1.0
builtin-interfaces==1.2.1
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chrome-gnome-shell==0.0.0
click==8.1.8
cmd_publisher==0.0.0
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.2.1
colcon-clean==0.2.1
colcon-cmake==0.2.29
colcon-common-extensions==0.3.0
colcon-core==0.19.0
colcon-defaults==0.2.9
colcon-devtools==0.3.0
colcon-installed-package-information==0.2.1
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-mixin==0.2.3
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.4.6
command-not-found==0.3
composition-interfaces==1.2.1
configobj==5.0.9
cov-core==1.15.0
coverage==7.6.12
cryptography==44.0.2
cupshelpers==1.0
cv-bridge==3.2.1
cycler==0.12.1
dbus-python==1.3.2
decorator==5.2.1
defer==1.0.6
demo-nodes-py==0.20.5
Deprecated==1.2.18
diagnostic-msgs==4.8.0
distlib==0.3.9
distro==1.9.0
distro-info==1.1+ubuntu0.2
docutils==0.21.2
domain-coordinator==0.10.0
dulwich==0.22.8
duplicity==0.8.21
empy==3.3.2
Encoder==1.1
encoders_node==0.0.0
evdev==1.9.1
example-interfaces==0.9.3
examples-rclpy-executors==0.15.3
examples-rclpy-minimal-action-client==0.15.3
examples-rclpy-minimal-action-server==0.15.3
examples-rclpy-minimal-client==0.15.3
examples-rclpy-minimal-publisher==0.15.3
examples-rclpy-minimal-service==0.15.3
examples-rclpy-minimal-subscriber==0.15.3
exceptiongroup==1.2.2
fastbencode==0.3.1
fasteners==0.19
fastimport==0.9.14
feedback_publisher==0.0.0
filelock==3.17.0
flake8==7.1.2
flake8-blind-except==0.2.1
flake8-builtins==2.5.0
flake8-class-newline==1.6.0
flake8-comprehensions==3.16.0
flake8-deprecated==2.2.1
flake8-docstrings==1.7.0
flake8-import-order==0.18.2
flake8-quotes==3.4.0
fonttools==4.56.0
fs==2.4.16
future==1.0.0
gast==0.6.0
geometry-msgs==4.8.0
gpg==1.16.0
html5lib==1.1
httplib2==0.22.0
idna==3.10
image-geometry==3.2.1
importlib_metadata==8.6.1
iniconfig==2.0.0
interactive-markers==2.3.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
keyring==25.6.0
kiwisolver==1.4.8
language-selector==0.1
lark==1.2.2
laser-geometry==2.4.0
launch==1.0.8
launch-ros==0.19.9
launch-testing==1.0.8
launch-testing-ros==0.19.9
launch-xml==1.0.8
launch-yaml==1.0.8
launchpadlib==2.1.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.7
lgpio==0.2.2.0
lifecycle-msgs==1.2.1
lockfile==0.12.2
logging-demo==0.20.5
louis==3.20.0
lqr_controller==0.0.0
lxml==4.8.0
lz4==4.4.3
macaroonbakery==1.3.4
Mako==1.3.9
map-msgs==2.1.0
MarkupSafe==3.0.2
matplotlib==3.5.1
mccabe==0.7.0
mercurial==6.9.3
message-filters==4.3.7
monotonic==1.6
more-itertools==10.6.0
motor_controller==0.0.0
mpi4py==4.0.3
mpmath==1.3.0
mpu6050_reader==0.0.0
nav-msgs==4.8.0
netifaces==0.11.0
nose2==0.15.1
notify2==0.3.1
numpy==1.21.5
oauthlib==3.2.2
olefile==0.47
osrf-pycommon==2.1.6
packaging==24.2
paramiko==3.5.1
pathspec==0.12.1
patiencediff==0.2.15
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pexpect==4.9.0
pigpio==1.78
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
protobuf==6.30.0
psutil==7.0.0
ptyprocess==0.7.0
py==1.11.0
pycairo==1.20.1
pycodestyle==2.12.1
pycparser==2.22
pycups==2.0.1
pydocstyle==6.3.0
pydot==3.0.4
pyflakes==3.2.0
PyGithub==2.6.1
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.10.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pynput==1.8.0
pyparsing==3.2.1
PyQt5==5.15.6
PyQt5_sip==12.17.0
pyRFC3339==2.0.1
pytest==8.3.5
pytest-cov==6.0.0
pytest-repeat==0.9.3
pytest-rerunfailures==15.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==1.0.0
python-gitlab==5.6.0
python-qt-binding==1.1.2
python-xlib==0.33
pythran==0.17.0
pytz==2025.1
pyxdg==0.28
PyYAML==6.0.2
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.16
rcutils==5.1.6
reportlab==3.6.8
requests==2.32.3
requests-toolbelt==1.0.0
resource-retriever==3.1.3
rmw-dds-common==1.6.0
roman==5.0
ros2action==0.18.12
ros2bag==0.15.14
ros2cli==0.18.12
ros2component==0.18.12
ros2doctor==0.18.12
ros2interface==0.18.12
ros2launch==0.19.9
ros2lifecycle==0.18.12
ros2multicast==0.18.12
ros2node==0.18.12
ros2param==0.18.12
ros2pkg==0.18.12
ros2run==0.18.12
ros2service==0.18.12
ros2topic==0.18.12
rosbag2-interfaces==0.15.14
rosbag2-py==0.15.14
rosdep==0.25.1
rosdep-modules==0.25.1
rosdistro==1.0.1
rosdistro-modules==1.0.1
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.6
rosidl-cli==3.1.6
rosidl-cmake==3.1.6
rosidl-generator-c==3.1.6
rosidl-generator-cpp==3.1.6
rosidl-generator-py==0.14.4
rosidl-parser==3.1.6
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.6
rosidl-typesupport-introspection-cpp==3.1.6
rospkg-modules==1.6.0
rpi-lgpio==0.6
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.4
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scantree==0.0.4
scipy==1.8.0
SecretStorage==3.3.3
sensor-msgs==4.8.0
sensor-msgs-py==4.8.0
shape-msgs==4.8.0
six==1.17.0
smbus2==0.5.0
snowballstemmer==2.2.0
soupsieve==2.6
sros2==0.10.6
ssh-import-id==5.11
statistics-msgs==1.2.1
std-msgs==4.8.0
std-srvs==4.8.0
stereo-msgs==4.8.0
sympy==1.9
systemd-python==234
teleop-twist-keyboard==2.4.0
tf2-geometry-msgs==0.25.12
tf2-kdl==0.25.12
tf2-msgs==0.25.12
tf2-py==0.25.12
tf2-ros-py==0.25.12
tf2-tools==0.25.12
toml==0.10.2
tomli==2.2.1
topic-monitor==0.20.5
trajectory-msgs==4.8.0
turtlesim==1.4.2
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.17.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==16.0.0
unique-identifier-msgs==2.2.1
urllib3==2.3.0
vcstool==0.3.0
vcstools==0.1.42
virtualenv==20.29.3
visualization-msgs==4.8.0
wadllib==2.0.0
webencodings==0.5.1
wrapt==1.17.2
xdg==6.0.0
xkit==0.0.0
zipp==3.21.0
